+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/post/repository/PostRepository.kt
@@ -10,12 +10,12 @@ import org.springframework.stereotype.Repository
 import siksha.wafflestudio.core.domain.post.data.Post
 
 @Repository
-interface PostRepository: JpaRepository<Post, Long> {
+interface PostRepository: JpaRepository<Post, Int> {
     @EntityGraph(attributePaths = ["user"])
     @Query("SELECT p FROM post p WHERE p.board.id = :boardId ORDER BY p.createdAt DESC")
-    fun findPageByBoardId(@Param("boardId") boardId: Long, pageable: Pageable): Page<Post>
+    fun findPageByBoardId(@Param("boardId") boardId: Int, pageable: Pageable): Page<Post>
 
     @EntityGraph(attributePaths = ["user"])
     @Query("SELECT p FROM post p WHERE p.user.id = :userId ORDER BY p.createdAt DESC")
-    fun findPageByUserId(@Param("userId") userId: Long, pageable: Pageable): Page<Post>
+    fun findPageByUserId(@Param("userId") userId: Int, pageable: Pageable): Page<Post>
 }
